+++ packages/react-pdf-toolkit/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @ballerine/react-pdf-toolkit
 
+## 1.2.34
+
+### Patch Changes
+
+- Updated dependencies
+- Bump
+- Updated dependencies
+  - @ballerine/config@1.1.18
+  - @ballerine/ui@0.5.34
+
 ## 1.2.33
 
 ### Patch Changes